🛠 #307 DanggnGameController로 변경
- 기존 Channel을 사용한 DanggnShaker 방식은 가독성도 안좋고 의존성이 너무 이상하게 걸려있어서 새로운 구조로 수정하려고 합니다.
